Work on importer before first round of testing close #4752, #4765, #4766, #4767, #4768

+++ components/insight/SRC/org/openmicroscopy/shoola/util/ui/UIUtilities.java
@@ -1507,13 +1507,13 @@ public class UIUtilities
 	 * @param v The value to convert.
 	 * @return See above.
 	 */
-	public static String formatFileSize(long v)
+	public static String formatFileSize(double v)
 	{
 		if (v < 0) return "";
 
 		//if (v < 1000) 
 		//	return NumberFormat.getInstance().format(v)+" b";
-		long value = v;///1000;
+		double value = v;///1000;
 		if (value <= 1000) 
 			return NumberFormat.getInstance().format(value)+" Kb";
 		value = value/1000;
